pyproject.toml
[tool.poetry]
name = "task-tracker-kanban"
version = "0.1.0"
description = "Трекер задач с канбаном"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.27"}
passlib = "^1.7.4"
python-jose = {extras = ["pycryptodome"], version = "^3.3.0"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
